progress: and away goes those lines
