+++ jac/examples/plugins/cmd_show/cmd_show/show.py
@@ -1,7 +1,8 @@
 import os
 
 from jaclang.cli.cmdreg import cmd_registry
-from jaclang.runtimelib.plugin.default import hookimpl
+
+# from jaclang.runtimelib.plugin.default import hookimpl
 
 import pygments
 from pygments.formatters import TerminalFormatter
@@ -15,7 +16,7 @@ class JacCmd:
     """Jac CLI."""
 
     @staticmethod
-    @hookimpl
+    # @hookimpl
     def create_cmd() -> None:
         """Creating Jac CLI cmds."""
