+++ jac/jaclang/compiler/passes/tool/tests/fixtures/corelib_fmt.jac
@@ -3,7 +3,7 @@
 import:py from datetime, datetime;
 import:py from uuid, UUID, uuid4;
 import:py from jaclang.compiler.constant, EdgeDir;
-import:py from jaclang.runtimelib.plugin, Architype, ArchitypeProtocol, DSFunc, AbsRootHook, hookimpl;
+import:py from jaclang.runtimelib.machine, Architype, ArchitypeProtocol, DSFunc, AbsRootHook;
 
 enum AccessMode;
