+++ jac/jaclang/runtimelib/__init__.py
@@ -1,7 +1,1 @@
 """Core primitives for Jaseci."""
-
-from jaclang.runtimelib.plugin.default import JacFeatureImpl
-from jaclang.runtimelib.plugin.feature import plugin_manager
-
-plugin_manager.register(JacFeatureImpl)
-plugin_manager.load_setuptools_entrypoints("jac")
